Initial architecture
Shows the initial architecture of the worshipjs project. The 3-axis orientation modifier has been added to serve as an example for the rest.

TODO: modified the architecture so that this is one element that exists on a page (rather than all connected, is is presently, to the root element of the app)